Update invoice extraction to standardized schema
