+++ samples/assets/invoices/invoice_5.json
@@ -1,50 +1,63 @@
 {
-    "fname": "invoice_5.pdf",
-    "expected": {
-        "customer_name": "ALD AUTOMOTIVE PVT LTD",
-        "customer_address": null,
-        "customer_tax_id": "27AAFCA0924K1ZR",
-        "shipping_address": null,
-        "purchase_order": null,
-        "invoice_id": "GT\\5153",
-        "invoice_date": "2019-02-27",
-        "payable_by": null,
-        "vendor_name": "DILIP TYRES",
-        "vendor_address": {
-            "street": "Pushparaj Chowk",
-            "city": "Sangli",
-            "state": "Maharashtra",
-            "postal_code": null,
-            "country": null
+  "fname": "invoice_5.pdf",
+  "0_expected": {
+    "customer_name": "ALD AUTOMOTIVE PVT LTD",
+    "customer_tax_id": "27AAFCA0924K1ZR",
+    "customer_address": {
+      "street": null,
+      "city": null,
+      "state": "Maharashtra",
+      "postal_code": null,
+      "country": "India"
+    },
+    "shipping_address": null,
+    "purchase_order": null,
+    "invoice_id": "GT\\5153",
+    "invoice_date": "2019-02-27",
+    "due_date": null,
+    "vendor_name": "DILIP TYRES",
+    "vendor_address": {
+      "street": "Pushparaj Chowk",
+      "city": "Sangli",
+      "state": "Maharashtra",
+      "postal_code": null,
+      "country": "India"
+    },
+    "vendor_tax_id": "27AAEPH8682F1ZI",
+    "remittance_address": null,
+    "subtotal": {
+      "currency_code": "INR",
+      "amount": 2287.5
+    },
+    "total_discount": null,
+    "total_tax": {
+      "currency_code": "INR",
+      "amount": 640.5
+    },
+    "invoice_total": {
+      "currency_code": "INR",
+      "amount": 2928.0
+    },
+    "payment_term": null,
+    "items": [
+      {
+        "product_code": "40111090",
+        "description": "165.80R/14 ASSURANCE DP T/L TYRES",
+        "quantity": 1,
+        "tax": {
+          "currency_code": "INR",
+          "amount": 640.5
         },
-        "vendor_tax_id": "27AAEPH8682F1ZI",
-        "remittance_address": null,
-        "subtotal": 2287.50,
-        "total_discount": null,
-        "total_tax": 640.50,
-        "invoice_total": 2928.00,
-        "payment_terms": null,
-        "items": [
-            {
-                "product_code": "40111090",
-                "description": "165.80R/14 ASSURANCE DP T/L TYRES",
-                "quantity": 1,
-                "tax": 640.50,
-                "tax_rate": "14%",
-                "unit_price": 2287.50,
-                "total": 2928.00,
-                "reason": null
-            }
-        ],
-        "total_item_quantity": 1,
-        "items_customer_signature": null,
-        "items_vendor_signature": {
-            "signatory": null,
-            "is_signed": true
+        "unit_price": {
+          "currency_code": "INR",
+          "amount": 2287.5
         },
-        "returns": null,
-        "total_return_quantity": null,
-        "returns_customer_signature": null,
-        "returns_vendor_signature": null
-    }
+        "total": {
+          "currency_code": "INR",
+          "amount": 2928.0
+        }
+      }
+    ],
+    "customer_signature": null
+  }
 }
