+++ samples/python/modules/samples/evaluation/comparison.py
@@ -1,7 +1,5 @@
-from typing import Optional
 import pandas as pd
 
-from samples.models.classification import Classifications
 from samples.utils.value_utils import flatten_dict
 
 
@@ -29,11 +27,13 @@ def get_extraction_comparison(expected: dict, actual: dict, confidence: dict, ac
 
     rows = []
     for key in all_keys:
+        key_confidence = confidence_flat.get(f"{key}_confidence", None)
+
         rows.append({
             "Field": key,
             "Expected": expected_flat.get(key),
             "Extracted": extracted_flat.get(key),
-            "Confidence": f"{confidence_flat.get(f"{key}_confidence", 0.0) * 100:.2f}%",
+            "Confidence": f"{key_confidence * 100:.2f}%" if key_confidence else 'N/A',
             "Accuracy": f"{'Match' if accuracy_flat.get(f"{key}", 0.0) == 1.0 else 'Mismatch'}"
         })
     df = pd.DataFrame(rows)
@@ -43,30 +43,3 @@ def get_extraction_comparison(expected: dict, actual: dict, confidence: dict, ac
 
     df = df.style.apply(highlight_row, axis=1)
     return df
-
-
-def get_classification_comparison(expected: Classifications, extracted: Classifications, confidence: Optional[dict] = None):
-    if confidence is not None:
-        confidence_flat = flatten_dict(confidence)
-
-    rows = []
-    for classification in expected.classifications:
-        extracted_classification = extracted.get_classification(
-            classification.page_number)
-        row = {
-            "Page": classification.page_number,
-            "Expected": classification.classification,
-            "Extracted": extracted_classification.classification
-        }
-
-        # If extracted_classification is a Classification object, add extra metadata
-        if hasattr(extracted_classification, 'similarity'):
-            row["Similarity"] = extracted_classification.similarity
-
-        if confidence is not None:
-            row["Confidence"] = confidence_flat.get(
-                f"classifications_{classification.page_number}_classification_confidence", None)
-
-        rows.append(row)
-    df = pd.DataFrame(rows)
-    return df.to_markdown(index=False, tablefmt="unsafehtml")
